+++ folding-schemes/src/folding/mod.rs
@@ -1,6 +1,5 @@
 pub mod circuits;
 pub mod hypernova;
-pub mod mova;
 pub mod nova;
 pub mod protogalaxy;
 pub mod traits;
